+++ IBLWorkFlow/Data/IBLUserListRepository.h
@@ -27,9 +27,9 @@
 /// 所选优惠条目ID
 @property (nonatomic, copy) NSString *discountItems;
 /// 总金额，以分为单位（必选）
-@property (nonatomic, assign) NSInteger totalCost;
+@property (nonatomic, assign) CGFloat totalCost;
 /// 优惠金额，以分为单位（可选）
-@property (nonatomic, assign) NSInteger preCost;
+@property (nonatomic, assign) CGFloat preCost;
 /// 其它金额，以分为单位（可选）
 @property (nonatomic, copy) NSString *otherCost;
 /// 临时赠送量
@@ -145,9 +145,9 @@
 //所选优惠条目ID，格式为以逗号为分割的ID字符串，例如所选优惠条目的ID为14000111111111和14000111111112，那么传递参数应为” 14000111111111，14000111111112” （可选）
 @property(nonatomic, copy) NSString *discountItems;
 //支付金额，以分为单位（必选）totalCost = 销售品金额+其它金额-优惠金额
-@property(nonatomic, assign) NSInteger totalCost;
+@property(nonatomic, assign) CGFloat totalCost;
 //优惠金额，以分为单位（可选）
-@property(nonatomic, assign) NSInteger preCost;
+@property(nonatomic, assign) CGFloat preCost;
 
 //其它金额，以分为单位（可选）
 @property(nonatomic, assign) NSInteger otherCost;
